A couple more TypeScript ports
